[TASK] Move country fetching to separate CountrySelectViewHelper and use it in BE as well
