+++ Classes/Controller/Application/PersonalDataController.php
@@ -3,7 +3,6 @@ namespace PAGEmachine\Ats\Controller\Application;
 
 use PAGEmachine\Ats\Domain\Model\ApplicationB;
 use PAGEmachine\Ats\Domain\Repository\CountryRepository;
-use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
 
 /*
  * This file is part of the PAGEmachine ATS project.
@@ -49,9 +48,6 @@ class PersonalDataController extends AbstractApplicationController
             $this->view->assign('defaultNationality', $this->countryRepository->findOneByIsoCodeA3($this->settings['defaultNationality']));
         }
 
-        $countries = $this->getStaticCountries();
-
-        $this->view->assign('countries', $countries);
         $this->view->assign("application", $application);
     }
 
@@ -66,29 +62,5 @@ class PersonalDataController extends AbstractApplicationController
     {
         $this->repository->addOrUpdate($application);
         $this->forward("editQualifications", "Application\\Qualifications", null, ['application' => $application->getUid()]);
-    }
-
-    protected function getStaticCountries()
-    {
-        if (!empty($this->settings['allowedStaticCountries'])) {
-            return $this->countryRepository->findCountriesByUids(
-                explode(',', $this->settings['allowedStaticCountries'])
-            );
-        } elseif (!empty($this->getStaticInfoTablesSettings()['countriesAllowed'])) {
-            return $this->countryRepository->findCountriesByISO3(
-                explode(',', $this->getStaticInfoTablesSettings()['countriesAllowed'])
-            );
-        }
-        return $this->countryRepository->findAll();
-    }
-
-    protected function getStaticInfoTablesSettings()
-    {
-        $configurationManager = $this->objectManager->get(ConfigurationManagerInterface::class);
-        return $configurationManager->getConfiguration(
-            ConfigurationManagerInterface::CONFIGURATION_TYPE_SETTINGS,
-            'StaticInfoTables',
-            'pi1'
-        );
     }
 }
