+++ Resources/Private/Partials/Application/Field/SelectCountry.html
@@ -1,3 +1,4 @@
+{namespace ats=PAGEmachine\Ats\ViewHelpers}
 <div class="form-group">
 	<div class="col-sm-3">
 		<f:render partial="Application/Field/Label" arguments="{fieldName : fieldName, translationKey : translationKey}" />
@@ -5,10 +6,10 @@
 	<div class="col-sm-9">
 		<f:if condition="{defaultCountry}">
 			<f:then>
-				<f:form.select property="{fieldName}" optionLabelField="localizedName" optionValueField="uid" options="{countries}" class="form-control" prependOptionLabel="{defaultCountry.localizedName}" prependOptionValue="{defaultCountry.uid}" />
+				<ats:form.countrySelect property="{fieldName}" optionLabelField="localizedName" optionValueField="uid" allowedStaticCountries="{settings.allowedStaticCountries}" class="form-control" prependOptionLabel="{defaultCountry.localizedName}" prependOptionValue="{defaultCountry.uid}" />
 			</f:then>
 			<f:else>
-				<f:form.select property="{fieldName}" optionLabelField="localizedName" optionValueField="uid" options="{countries}" class="form-control" />
+				<ats:form.countrySelect property="{fieldName}" optionLabelField="localizedName" optionValueField="uid" allowedStaticCountries="{settings.allowedStaticCountries}" class="form-control" />
 			</f:else>
 		</f:if>
 	</div>
