+++ Resources/Private/Partials/Backend/Application/Field/SelectNationality.html
@@ -1,4 +1,4 @@
-{namespace s=SJBR\StaticInfoTables\ViewHelpers}
+{namespace ats=PAGEmachine\Ats\ViewHelpers}
 <tr class="form-element">
 	<td>
 		<label>
@@ -6,8 +6,7 @@
 		</label>
 	</td>
 	<td>
-			<s:form.select name="staticInfoTablesCountry" property="{fieldName}" staticInfoTable="country" optionLabelField="shortNameEn" optionValueField="isoCodeA3" options="{}" class="form-control" defaultOptionValue="" defaultOptionLabel="{f:translate(key:'tx_ats.label.select.undefined')}" />
-		</div>
+		<ats:form.countrySelect property="{fieldName}" optionLabelField="localizedName" optionValueField="cn_iso_3" class="form-control" prependOptionValue="" prependOptionLabel="{f:translate(key:'tx_ats.label.select.undefined')}" />
 	</td>
 </tr>
